backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
websockets==12.0
python-dotenv==1.0.1
dnspython==2.4.2